backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
supabase
python-jose
